refactor(brett): board-boot.ts → board-replay + board-moderation-ui (T000661)
